employee_activity_tracker/requirements.txt
mysql-connector-python>=8.0
pymysql>=1.0
python-dotenv>=1.0
faker>=18.0
openai>=1.0
cryptography>=3.4